requirements.txt
backcall==0.1.0
decorator==4.4.1
Django==2.0.1
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.2
ptyprocess==0.6.0
pycryptodome==3.6.6
Pygments==2.5.2
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.1.8
